egps: fix lon/lat channel swap and float32 precision in egpsb decode
payload[0::6] encodes lon-delta and payload[1::6] encodes lat-delta
(firmware gps/mod.rs), but the decoder had them swapped, adding the
lon-delta to the lat reference and vice versa. This caused the per-sample
track to show E-W motion as N-S and vice versa.

Also fix float32 precision loss: data['lat'] and data['lon'] are parsed
from JSON as Python ints, so float32 + int stays float32. Float32 ULP
near 6e8 (deg*1e7) is ~64 units (~0.7m), collapsing nearby samples onto
the same grid point and creating staircase discontinuities. Cast the
references to float64 before adding the scaled offsets so the result is
float64 with ~0.03m resolution from the u16 encoding.

Also keep lat/lon as float64 in EgpsTimeseries.to_dataset() to preserve
that precision through to the xarray Dataset.

Add test_sfy4_01_egpsb_position: validates egpsb position packages from
SFY4-01 during a boat test on 2026-05-19 17:00-18:00 UTC (speeds up to
16 knots with stationary periods).

+++ sfy-processing/sfy/timeseries.py
@@ -543,14 +543,14 @@ class EgpsTimeseries:
                 }),
             'lat':
             xr.Variable(('time'),
-                        self.n.astype(np.float32),
+                        self.n.astype(np.float64),
                         attrs={
                             'unit': 'degrees_north',
                             'long_name': 'latitude',
                         }),
             'lon':
             xr.Variable(('time'),
-                        self.e.astype(np.float32),
+                        self.e.astype(np.float64),
                         attrs={
                             'unit': 'degrees_east',
                             'standard_name': 'longitude',
